@@ -14,12 +14,9 @@ class RandomSpellCommand extends Command
 {
     protected static $defaultName = 'app:random-spell';
     protected static $defaultDescription = 'Cast a random spell!';
-    private $logger;
 
-    public function __construct(LoggerInterface $logger)
+    public function __construct(private LoggerInterface $logger)
     {
-        $this->logger = $logger;
-
         parent::__construct();
     }
 
