@@ -49,7 +49,7 @@ final class QuestionFactory extends ModelFactory
                 true
             ),
             'askedAt' => self::faker()->dateTimeBetween('-100 days', '-1 minute'),
-            'votes' => rand(-20, 50),
+            'votes' => random_int(-20, 50),
             'owner' => UserFactory::new(),
         ];
     }
